[ch184] Moving Redux store into correct file structure
