+++ src/store/configureStore.js
@@ -1,0 +1,7 @@
+import { createStore } from "redux"
+import rootReducer from "../reducers/rootReducer"
+
+export default () => {
+  const store = createStore(rootReducer)
+  return store
+}
